requirements.txt
# Our favorite web framework
Django==1.9.6

# Read sensitive settings from the environment
python-decouple==3.0

# Ease path/files operations
Unipath==1.1

# 12fator-like database URL parser
dj-database-url==0.4.0

# HTTP server for WSGI
gunicorn==19.4.5
gevent==1.1.1

# Allow Heroku to serve static files
whitenoise==3.0
